# chore(ssa): Activate loop invariant code motion on ACIR functions