@vezenovm 2024-12-12
> I see a good enough improvement in regression_4709 alone that I'd say this is worth merging as-is

Yeah looks like it may be good to do on its own even without #6784